+++ rest/src/index.ts
@@ -1,9 +1,5 @@
 export * from './decorators';
-export * from './errorHandlers';
-export * from './interceptors';
 export * from './middlewares';
-export * from './rxjsOperators';
 
-export * from './rest.interface';
-export * from './tokens';
-export * from './types';
+export * from './misc/restHttpError.interface';
+