+++ src/errorHandlers/handle400WithValidationsFunc.ts
@@ -1,9 +1,11 @@
 import {HttpErrorResponse} from '@angular/common/http';
 import {Observable, of, throwError} from 'rxjs';
 
-import {Handle400WithValidationsOptions} from '../rest.interface';
-import {CLIENT_ERROR_NOTIFICATIONS, REST_CLIENT_ERRORS_RESPONSE_MAPPER, REST_CLIENT_VALIDATION_ERRORS_RESPONSE_MAPPER} from '../tokens';
-import {RestClientError, ClientValidationError} from '../types';
+import {Handle400WithValidationsOptions} from '../misc/httpError.interface';
+import {CLIENT_ERROR_NOTIFICATIONS, HTTP_CLIENT_ERROR_RESPONSE_MAPPER, HTTP_CLIENT_VALIDATION_ERROR_RESPONSE_MAPPER} from '../misc/tokens';
+import {RestClientError, ClientValidationError} from '../misc/httpErrors';
+
+//TODO: extract functionality and use it in interceptor
 
 /**
  * Handles http error response with code 400 and process validations errors and returns RestClientError or RestClientValidationError
@@ -15,8 +17,8 @@ export function handle400WithValidationsFunc(error: HttpErrorResponse, options: 
     //handles 400 code
     if(error.status == 400)
     {
-        const clientErrorsResponseMapper = options.clientErrorsResponseMapper ?? options.injector.get(REST_CLIENT_ERRORS_RESPONSE_MAPPER);
-        const clientValidationErrorsResponseMapper = options.clientValidationErrorsResponseMapper ?? options.injector.get(REST_CLIENT_VALIDATION_ERRORS_RESPONSE_MAPPER);
+        const clientErrorsResponseMapper = options.clientErrorsResponseMapper ?? options.injector.get(HTTP_CLIENT_ERROR_RESPONSE_MAPPER);
+        const clientValidationErrorsResponseMapper = options.clientValidationErrorsResponseMapper ?? options.injector.get(HTTP_CLIENT_VALIDATION_ERROR_RESPONSE_MAPPER);
         const errors = clientErrorsResponseMapper(error);
         const validationErrors = clientValidationErrorsResponseMapper(error);
         const notifications = options.injector.get(CLIENT_ERROR_NOTIFICATIONS, null);