+++ src/interceptors/index.ts
@@ -1,1 +1,2 @@
 export * from './httpServerError.interceptor';
+export * from './httpClientError.interceptor';