+++ src/serverValidation/serverValidationValidator.directive.ts
@@ -5,6 +5,8 @@ import {filter} from 'rxjs/operators';
 
 import {ServerValidationService} from './serverValidation.service';
 
+//TODO: try to get form control name
+
 /**
  * Name of server validations as error
  */